requirements.txt
selenium==4.23.1
pytest==8.3.2
pytest-xdist==3.6.1
allure-pytest==2.13.5
webdriver-manager==4.0.2
